+++ test/test/indexTest.js
@@ -1,13 +1,13 @@
-if (typeof(exports) !== 'undefined') {
+if (typeof (exports) !== 'undefined') {
 	var chai = require('chai');
 	var sinonChai = require("sinon-chai");
 	var sinon = require('sinon');
 	chai.use(sinonChai);
 
-	var libpath = (process.env['FUNCTION_FLOW_COV'] ? 'cov/' : '');
+	var libpath = (process.env['FUNCTION_FLOW_COV'] ? 'test/coverage/lib/' : 'lib/');
 
 	// The Lib we want to test
-	var FunctionFlow = require('../../' + libpath + 'index.js');
+	var FunctionFlow = require('../../' + libpath + 'functionFlow.js');
 }
 
 chai.Assertion.includeStack = true; // defaults to false
@@ -39,7 +39,6 @@ describe('methods without return value should return the flow', function() {
 	});
 
 });
-
 describe('ending methods should return undefined', function() {
 
 	function testMethod(methodName) {
@@ -57,9 +56,10 @@ describe('method order', function() {
 	it('start without any step, calls only the given function', function() {
 		var flow = new FunctionFlow();
 		var flowDone = sinon.spy();
+
 		flow.now(flowDone);
 		expect(flowDone).to.be.calledOnce;
-		expect(flowDone).always.have.been.calledWithExactly([]);
+		expect(flowDone).always.have.been.calledWithExactly(undefined, undefined);
 	});
 
 	it('the done callback is not called, when a previous step does not call the callback', function() {
@@ -76,7 +76,7 @@ describe('method order', function() {
 		var flowDone = sinon.spy(function() {
 			expect(firstStep).to.be.calledOnce;
 			expect(flowDone).to.be.calledOnce;
-			expect(flowDone).always.have.been.calledWithExactly([{error : undefined, data : undefined}]);
+			expect(flowDone).always.have.been.calledWithExactly([undefined], [undefined]);
 			done();
 		});
 		var firstStep = sinon.spy(function(flow) {
@@ -84,6 +84,7 @@ describe('method order', function() {
 				flow.done();
 			});
 		});
+
 		flow.run(firstStep).now(flowDone);
 
 	});
@@ -92,17 +93,19 @@ describe('method order', function() {
 describe('errors', function() {
 	it('a step throws an error', function(done) {
 		var flow = new FunctionFlow();
-		var flowDone = sinon.spy(function(result) {
-			expect(result).to.have.length(1);
-			result.forEach(function(res) {
-				expect(res.error).to.be.an.instanceof(Error);
-				expect(res.data).to.be.undefined;
+		var flowDone = sinon.spy(function(error, data) {
+			expect(error).to.have.length(1);
+			error.forEach(function(err) {
+				expect(err).to.be.an.instanceof(Error);
+			});
+			expect(data).to.have.length(1);
+			data.forEach(function(dat) {
+				expect(dat).to.be.undefined;
 			});
 			expect(firstStep).to.be.calledOnce;
 			expect(flowDone).to.be.calledOnce;
 			done();
 		});
-
 		var firstStep = sinon.spy(function(flow) {
 			throw new ReferenceError('testError');
 			setTimeout(function() {
@@ -113,16 +116,19 @@ describe('errors', function() {
 		flow.run(firstStep).now(flowDone);
 	});
 
-	it('a error stops the execution of a flow', function(done) {
+	it('a returned error stops the execution of a flow', function(done) {
 		var flow = new FunctionFlow();
-		var flowDone = sinon.spy(function(result) {
-			expect(result).to.have.length(1);
-			result.forEach(function(res) {
-				expect(res.error).to.be.an.instanceof(Error);
-				expect(res.data).to.be.undefined;
+		var flowDone = sinon.spy(function(error, data) {
+			expect(error).to.have.length(1);
+			error.forEach(function(err) {
+				expect(err).to.be.instanceof(Error);
+			});
+			expect(data).to.have.length(1);
+			data.forEach(function(dat) {
+				expect(dat).to.be.undefined;
 			});
 			expect(firstStep).to.be.calledOnce;
-			expect(secondsStep).to.not.be.called;
+			expect(secondStep).to.not.be.called;
 			expect(flowDone).to.be.calledOnce;
 			done();
 		});
@@ -132,26 +138,28 @@ describe('errors', function() {
 				flow.done(new Error('TestError'), undefined);
 			});
 		});
-		var secondsStep = sinon.spy(function(flow) {
-			assert.fail();
+		var secondStep = sinon.spy(function(flow) {
 			setTimeout(function() {
 				flow.done(undefined, true);
 			});
-
 		});
 
-		flow.run(firstStep).run(secondsStep).now(flowDone);
+		flow.run(firstStep).run(secondStep).now(flowDone);
 	});
+
 	it('a thrown error stops the execution of a flow', function(done) {
 		var flow = new FunctionFlow();
-		var flowDone = sinon.spy(function(result) {
-			expect(result).to.have.length(1);
-			result.forEach(function(res) {
-				expect(res.error).to.be.an.instanceof(Error);
-				expect(res.data).to.be.undefined;
+		var flowDone = sinon.spy(function(error, data) {
+			expect(error).to.have.length(1);
+			error.forEach(function(err) {
+				expect(err).to.be.instanceof(Error);
+			});
+			expect(data).to.have.length(1);
+			data.forEach(function(dat) {
+				expect(dat).to.be.undefined;
 			});
 			expect(firstStep).to.be.calledOnce;
-			expect(secondsStep).to.not.be.called;
+			expect(secondStep).to.not.be.called;
 			expect(flowDone).to.be.calledOnce;
 			done();
 		});
@@ -162,26 +170,199 @@ describe('errors', function() {
 				flow.done(undefined, true);
 			});
 		});
-		var secondsStep = sinon.spy(function(flow) {
+		var secondStep = sinon.spy(function(flow) {
 			setTimeout(function() {
 				flow.done(undefined, true);
 			});
 		});
 
-		flow.run(firstStep).run(secondsStep).now(flowDone);
+		flow.run(firstStep).run(secondStep).now(flowDone);
 	});
 
+	describe('configuratable behavior of error handling', function() {
+		var flow;
+		var firstStepA;
+		var firstStepB;
+		var secondStep;
+
+		beforeEach(function() {
+			flow = new FunctionFlow();
+			firstStepA = sinon.stub().throws(new ReferenceError('TestError'));
+			firstStepB = sinon.spy(function(flow) {
+				setTimeout(function() {
+					flow.done(undefined, 'b');
+				});
+			});
+			secondStep = sinon.spy(function(flow) {
+				flow.done(undefined, '2');
+			});
+		});
+
+		describe('default, #onErrorStopAfterStep(), parallel tasks are still all executed, even when one has an error', function() {
+
+			var expections = function(error, data) {
+				expect(error).to.have.length(2);
+				expect(error[0]).to.be.instanceof(Error);
+				expect(error[1]).to.be.undefined;
+				expect(data).to.have.length(2);
+				expect(data[0]).to.be.undefined;
+				expect(data[1]).to.be.equal('b');
+				expect(firstStepA).to.be.calledOnce;
+				expect(firstStepB).to.be.calledOnce;
+				expect(secondStep).to.not.be.called;
+			};
+
+			it('not called, default value', function(done) {
+				var flowDone = sinon.spy(function(error, data) {
+					expections(error, data);
+					done();
+				});
+
+				flow.run(firstStepA).and(firstStepB).run(secondStep).now(flowDone);
+			});
+
+			it('after calling #onErrorStopAfterStep()', function(done) {
+				var flowDone = sinon.spy(function(error, data) {
+					expections(error, data);
+					done();
+				});
+
+				flow.onErrorStopAfterStep().run(firstStepA).and(firstStepB).run(secondStep).now(flowDone);
+			});
+		});
+
+		describe('#onErrorNeverStop()', function() {
+
+			var expections = function(error, data) {
+				expect(error).to.have.length(1);
+				expect(error[0]).to.be.undefined;
+				expect(data).to.have.length(1);
+				expect(data[0]).to.be.equal('2');
+				expect(firstStepA).to.be.calledOnce;
+				expect(firstStepB).to.be.calledOnce;
+				expect(secondStep).to.be.calledOnce;
+			};
+
+			it('an error doesn\'t stop the execution', function(done) {
+
+				var flowDone = sinon.spy(function(error, data) {
+					expections(error, data);
+					done();
+				});
+
+				flow.onErrorNeverStop().run(firstStepA).and(firstStepB).run(secondStep).now(flowDone);
+			});
+		});
+
+		describe('#onErrorStopASAP()', function() {
+
+			var errorFunc;
+			var errorFuncAsync;
+			var returningSync;
+			var returningAsync;
+
+			beforeEach(function() {
+				errorFunc = sinon.stub().throws('Demo');
+				errorFuncAsync = sinon.spy(function(flow) {
+					setTimeout(function() {
+						flow.done(new Error('Demo'));
+					});
+				});
+				returningSync = sinon.spy(function(flow) {
+					flow.done(undefined, 'Earth');
+				});
+				returningAsync = sinon.spy(function(flow) {
+					setTimeout(function() {
+						flow.done(undefined, 'Earth');
+					});
+				});
+			});
+
+			it('when an error appeared, no new parallel task is started', function(done) {
+				flow.onErrorStopASAP().run(returningSync).and(errorFunc).and(returningSync).now(function(error, data) {
+					expect(returningSync).to.be.calledOnce;
+					expect(errorFunc).to.be.calledOnce;
+					expect(error[0]).to.be.undefined;
+					expect(error[1]).to.be.instanceof(Error);
+					expect(data[0]).to.be.equal('Earth');
+					expect(data[1]).to.be.undefined;
+
+					done();
+				});
+			});
+
+			it('after an error occured the data and error arrays are not manipulated by already started tasks', function(done) {
+				flow.onErrorStopASAP().run(returningSync).and(errorFuncAsync).and(returningAsync).now(function(error, data) {
+					expect(returningSync).to.be.calledOnce;
+					expect(returningAsync).to.be.calledOnce;
+					expect(errorFuncAsync).to.be.calledOnce;
+					setTimeout(function() {
+						expect(error).to.have.length(2);
+						expect(data).to.have.length(2);
+						done();
+					});
+				});
+			});
+		});
+	});
 });
 
+describe('parsing arguments with ".with()"', function() {
+	it('parsing one argument', function() {
+		var flow = new FunctionFlow();
 
-describe('run for each with arguments', function() {
+		var firstStep = sinon.spy(function(flow, parsedArgumentA) {
+
+			expect(parsedArgumentA).to.be.equal('argumentTestA');
+			flow.done();
+		});
+		flow.run(firstStep).with('argumentTestA').now(function() {
+		});
+	});
+
+	it('parsing three argument', function() {
+		var flow = new FunctionFlow();
+
+		var firstStep = sinon.spy(function(flow, parsedArgumentA, b, c) {
+
+			expect(parsedArgumentA).to.be.equal('argumentTestA');
+			expect(b).to.be.equal(2);
+			expect(c).to.be.equal(3);
+			flow.done();
+		});
+		flow.run(firstStep).with('argumentTestA', 2, 3).now(function() {
+		});
+	});
+});
+
+describe('combination of .forEach() and .with()', function() {
+	it('.with() arguments should be parsed before the .forEach() arguments', function(done) {
+		var flow = new FunctionFlow();
+		var stub = sinon.stub();
+		flow.run(stub).with('First').forEach(['SecondA','SecondB']).now(function() {
+			expect(stub.getCall(0).args).to.have.length(3);
+			expect(stub.getCall(0).args[1]).to.be.equal('First');
+			expect(stub.getCall(0).args[2]).to.be.equal('SecondA');
+			expect(stub.getCall(1).args).to.have.length(3);
+			expect(stub.getCall(1).args[1]).to.be.equal('First');
+			expect(stub.getCall(1).args[2]).to.be.equal('SecondB');
+			done();
+		});
+		stub.yieldTo('done');
+	});
+});
+
+describe('using .forEach()', function() {
 	it('calling with one element, single argument', function(done) {
 		var flow = new FunctionFlow();
-		var flowDone = sinon.spy(function(result) {
-			expect(result).to.have.length(1);
-			result.forEach(function(res) {
-				expect(res.error).to.be.undefined;
-				expect(res.data).to.be.true;
+		var flowDone = sinon.spy(function(error, data) {
+			expect(error).to.have.length(1);
+			error.forEach(function(err) {
+				expect(err).to.be.undefined;
+			});
+			expect(data).to.have.length(1);
+			data.forEach(function(dat) {
+				expect(dat).to.be.true;
 			});
 			expect(firstStep).to.be.calledOnce;
 			done();
@@ -194,41 +375,63 @@ describe('run for each with arguments', function() {
 				flow.done(undefined, true);
 			});
 		});
-		flow.run(firstStep).forEach([['argumentTestA']]).now(flowDone);
-
+		flow.run(firstStep).forEach(['argumentTestA']).now(flowDone);
 	});
-	it('calling with one element, two argument', function(done) {
+
+	it('order of function calls with .forEach()', function(done) {
+		var spy = sinon.spy(function(flow, nr) {
+			expect(spy.callCount).to.be.equal(nr);
+			flow.done(undefined, nr);
+		});
+		
 		var flow = new FunctionFlow();
-		var flowDone = sinon.spy(function(result) {
-			expect(result).to.have.length(1);
-			result.forEach(function(res) {
-				expect(res.error).to.be.undefined;
-				expect(res.data).to.be.true;
+		var arr = [1, 2, 3, 4, 5];
+		
+		flow.run(spy).forEach(arr).now(function(error, data) {
+			error.forEach(function(value) {
+				expect(value).to.be.undefined;
 			});
-			expect(firstStep).to.be.calledOnce;
+			expect(arr).to.be.deep.equal(arr);
+			expect(spy.callCount).to.be.equal(5);
 			done();
 		});
+	});
 
-		var refTest = {A : "B"};
-
-		var firstStep = sinon.spy(function(flow, parsedArgumentA, parsedArgumentB) {
-
-			expect(parsedArgumentA).to.be.equal('argumentTestA');
-			expect(parsedArgumentB).to.be.equal(refTest);
-			setTimeout(function() {
-				flow.done(undefined, true);
+	describe('Array parsed to .forEach() is not destroyed', function() {
+		it('without nested arguments', function(done) {
+			var flow = new FunctionFlow();
+			var spy = sinon.stub();
+			var arr = [1, 2, 3, 4, 5];
+			flow.run(spy).forEach(arr).now(function() {
+				expect(spy.callCount).to.be.equal(5);
+				expect(arr).to.be.deep.equal([1, 2, 3, 4, 5]);
+				done();
 			});
+			spy.yieldTo('done');
 		});
-		flow.run(firstStep).forEach([['argumentTestA', refTest]]).now(flowDone);
+		it('with nested arguments', function(done) {
+			var flow = new FunctionFlow();
+			var spy = sinon.stub();
+			var arr = [[1], [2], [3], [4], [5]];
+			flow.run(spy).forEach(arr).now(function() {
+				expect(spy.callCount).to.be.equal(5);
+				expect(arr).to.be.deep.equal([[1], [2], [3], [4], [5]]);
+				done();
+			});
+			spy.yieldTo('done');
+		});
 	});
 
 	it('calling with two element, single argument', function(done) {
 		var flow = new FunctionFlow();
-		var flowDone = sinon.spy(function(result) {
-			expect(result).to.have.length(2);
-			result.forEach(function(res) {
-				expect(res.error).to.be.undefined;
-				expect(res.data).to.be.true;
+		var flowDone = sinon.spy(function(error, data) {
+			expect(error).to.have.length(2);
+			error.forEach(function(err) {
+				expect(err).to.be.undefined;
+			});
+			expect(data).to.have.length(2);
+			data.forEach(function(dat) {
+				expect(dat).to.be.true;
 			});
 			expect(firstStep).to.be.calledTwice;
 			done();
@@ -258,14 +461,17 @@ describe('run for each with arguments', function() {
 describe('data parsing to next step', function() {
 	it('second step gets data of first step', function(done) {
 		var flow = new FunctionFlow();
-		var flowDone = sinon.spy(function(result) {
-			expect(result).to.have.length(1);
-			result.forEach(function(res) {
-				expect(res.error).to.be.undefined;
-				expect(res.data).to.be.true;
+		var flowDone = sinon.spy(function(error, data) {
+			expect(error).to.have.length(1);
+			error.forEach(function(err) {
+				expect(err).to.be.undefined;
+			});
+			expect(data).to.have.length(1);
+			data.forEach(function(dat) {
+				expect(dat).to.be.true;
 			});
 			expect(firstStep).to.be.calledOnce;
-			expect(secondsStep).to.be.calledOnce;
+			expect(secondStep).to.be.calledOnce;
 			expect(flowDone).to.be.calledOnce;
 			done();
 		});
@@ -275,37 +481,43 @@ describe('data parsing to next step', function() {
 				flow.done(undefined, false);
 			});
 		});
-
-		var secondsStep = sinon.spy(function(flow) {
-			expect(flow.previousStepResult).to.have.length(1);
-			flow.previousStepResult.forEach(function(res) {
-				expect(res.error).to.be.undefined;
-				expect(res.data).to.be.false;
+		var secondStep = sinon.spy(function(flow) {
+			expect(flow.previousStep.error).to.have.length(1);
+			flow.previousStep.error.forEach(function(err) {
+				expect(err).to.be.undefined;
+			});
+			expect(flow.previousStep.data).to.have.length(1);
+			flow.previousStep.data.forEach(function(dat) {
 			});
 			setTimeout(function() {
 				flow.done(undefined, true);
 			});
 		});
 
-		flow.run(firstStep).then(secondsStep).now(flowDone);
+		flow.run(firstStep).run(secondStep).now(flowDone);
 	});
+
+
+
 });
 
 describe('data collecting', function() {
 	describe('callback for start sums up data', function() {
 		it('single step is giving back a boolean false', function(done) {
 			var flow = new FunctionFlow();
-			var flowDone = sinon.spy(function(result) {
-				expect(result).to.have.length(1);
-				result.forEach(function(res) {
-					expect(res.error).to.be.undefined;
-					expect(res.data).to.be.false;
+			var flowDone = sinon.spy(function(error, data) {
+				expect(error).to.have.length(1);
+				error.forEach(function(err) {
+					expect(err).to.be.undefined;
+				});
+				expect(data).to.have.length(1);
+				data.forEach(function(dat) {
+					expect(dat).to.be.false;
 				});
 				expect(firstStep).to.be.calledOnce;
 				expect(flowDone).to.be.calledOnce;
 				done();
 			});
-
 			var firstStep = sinon.spy(function(flow) {
 				setTimeout(function() {
 					flow.done(undefined, false);
@@ -317,11 +529,14 @@ describe('data collecting', function() {
 
 		it('single step with two parallel run', function(done) {
 			var flow = new FunctionFlow();
-			var flowDone = sinon.spy(function(result) {
-				expect(result).to.have.length(2);
-				result.forEach(function(res) {
-					expect(res.error).to.be.undefined;
-					expect(res.data).to.be.false;
+			var flowDone = sinon.spy(function(error, data) {
+				expect(error).to.have.length(2);
+				error.forEach(function(err) {
+					expect(err).to.be.undefined;
+				});
+				expect(data).to.have.length(2);
+				data.forEach(function(dat) {
+					expect(dat).to.be.false;
 				});
 				expect(firstStep).to.be.calledTwice;
 				expect(flowDone).to.be.calledOnce;
@@ -336,50 +551,18 @@ describe('data collecting', function() {
 			flow.run(firstStep).and(firstStep).now(flowDone);
 		});
 
-		it('two steps, first with parallel tasks', function(done) {
-			var flow = new FunctionFlow();
-			var flowDone = sinon.spy(function(result) {
-				expect(result).to.have.length(1);
-				result.forEach(function(res) {
-					expect(res.error).to.be.undefined;
-					expect(res.data).to.be.false;
-				});
-				expect(secondStepCollecting).to.be.calledOnce;
-				expect(flowDone).to.be.calledOnce;
-				done();
-			});
-			var firstStep = sinon.spy(function(flow) {
-				setTimeout(function() {
-					flow.done(undefined, true);
-				});
-			});
-
-			var secondStepCollecting = sinon.spy(function(flow) {
-				expect(firstStep).to.be.calledTwice;
-				expect(flow.previousStepResult).to.be.deep.equal([
-					{error : undefined, data : true},
-					{error : undefined, data : true}
-				]);
-				setTimeout(function() {
-					flow.done(undefined, false);
-				});
-			});
-
-			flow.run(firstStep).and(firstStep).then(secondStepCollecting).now(flowDone);
-		});
-
 		describe('result order in parallel run', function() {
 
 			it('first attached function is faster', function(done) {
 				var flow = new FunctionFlow();
-				var flowDone = sinon.spy(function(result) {
-					expect(result).to.have.length(2);
-					result.forEach(function(res) {
-						expect(res.error).to.be.undefined;
+				var flowDone = sinon.spy(function(error, data) {
+					expect(error).to.have.length(2);
+					error.forEach(function(err) {
+						expect(err).to.be.undefined;
 					});
-
-					expect(result[0].data).to.be.false;
-					expect(result[1].data).to.be.true;
+					expect(data).to.have.length(2);
+					expect(data[0]).to.be.false;
+					expect(data[1]).to.be.true;
 
 					expect(firstRun).to.be.calledOnce;
 					expect(secondRun).to.be.calledOnce;
@@ -402,13 +585,14 @@ describe('data collecting', function() {
 
 			it('second attached function is faster', function(done) {
 				var flow = new FunctionFlow();
-				var flowDone = sinon.spy(function(result) {
-					expect(result).to.have.length(2);
-					result.forEach(function(res) {
-						expect(res.error).to.be.undefined;
+				var flowDone = sinon.spy(function(error, data) {
+					expect(error).to.have.length(2);
+					error.forEach(function(err) {
+						expect(err).to.be.undefined;
 					});
-					expect(result[0].data).to.be.false;
-					expect(result[1].data).to.be.true;
+					expect(data).to.have.length(2);
+					expect(data[0]).to.be.false;
+					expect(data[1]).to.be.true;
 
 					expect(firstRun).to.be.calledOnce;
 					expect(secondRun).to.be.calledOnce;
@@ -421,7 +605,9 @@ describe('data collecting', function() {
 					}, 50);
 				});
 				var secondRun = sinon.spy(function(flow) {
-					flow.done(undefined, true);
+					setTimeout(function() {
+						flow.done(undefined, true);
+					});
 				});
 
 				flow.run(firstRun).and(secondRun).now(flowDone);