+++ history.md
@@ -1,4 +1,13 @@
 
+0.0.2 / 2013-12-31
+==================
+
+  * added .forEach() to run a function for an array of arguments
+  * renamed .do() to .run()
+  * added .with() to specify arguments for a step
+  * onError behaviour is now settable with methods: .onErrorStopASAP(), .onErrorStopAfterStep(), .onErrorNeverStop()
+  * added some examples
+
 0.0.1 / 2013-09-29
 ==================